+++ feature/links/application/hooks/link/index.tsx
@@ -1,3 +1,4 @@
 export * from "./useLinkInput";
 export * from "./useLinks";
 export * from "./useLinkAction";
+export * from "./useLinksModals";
